+++ examples/kitchen-sink/pages/documentation/[...filename].tsx
@@ -26,12 +26,8 @@ export default function Home(
 }
 
 export const getStaticProps = async ({ params }) => {
-  const variables = { relativePath: `${params.filename}.md` }
-  let props = await client.queries.documentation(variables)
-
-  // if (process.env.VERCEL_ENV === 'preview') {
-  props = await expandWithMetadata(props, client)
-  // }
+  const variables = { relativePath: `${params.filename.join('/')}.md` }
+  const props = await client.queries.documentation(variables)
   return {
     props: { ...props, variables },
   }
@@ -41,7 +37,7 @@ export const getStaticPaths = async () => {
   const connection = await client.queries.documentationConnection()
   return {
     paths: connection.data.documentationConnection.edges.map((post) => ({
-      params: { filename: post.node._sys.filename },
+      params: { filename: post.node._sys.breadcrumbs },
     })),
     fallback: 'blocking',
   }