+++ experimental-examples/tina-cloud-starter-iframe/tina/config.tsx
@@ -48,7 +48,7 @@ const config = defineStaticConfig({
         format: "mdx",
         ui: {
           router: ({ document }) => {
-            return `/posts/${document._sys.filename}`;
+            return `/posts/${document._sys.breadcrumbs.join("/")}`;
           },
         },
         fields: [