+++ packages/tinacms/src/admin/components/GetCollection.tsx
@@ -16,6 +16,7 @@ export const useGetCollection = (
   cms: TinaCMS,
   collectionName: string,
   includeDocuments: boolean = true,
+  folder: { loading: boolean; fullyQualifiedName: string },
   after: string = '',
   sortKey?: string,
   filterArgs?: FilterArgs
@@ -32,7 +33,7 @@ export const useGetCollection = (
 
   useEffect(() => {
     const fetchCollection = async () => {
-      if (await api.isAuthenticated()) {
+      if ((await api.isAuthenticated()) && !folder.loading) {
         const { name, order } = JSON.parse(sortKey || '{}')
         const validSortKey = collectionExtra.fields
           ?.map((x) => x.name)
@@ -43,6 +44,7 @@ export const useGetCollection = (
           const collection = await api.fetchCollection(
             collectionName,
             includeDocuments,
+            filterArgs?.filterField ? '' : folder.fullyQualifiedName,
             after,
             validSortKey,
             order,
@@ -65,7 +67,15 @@ export const useGetCollection = (
     setLoading(true)
     fetchCollection()
     // TODO: useDebounce
-  }, [cms, collectionName, resetState, after, sortKey])
+  }, [
+    cms,
+    collectionName,
+    folder.loading,
+    folder.fullyQualifiedName,
+    resetState,
+    after,
+    sortKey,
+  ])
 
   const reFetchCollection = () => setResetSate((x) => x + 1)
 
@@ -75,6 +85,7 @@ export const useGetCollection = (
 const GetCollection = ({
   cms,
   collectionName,
+  folder,
   includeDocuments = true,
   startCursor,
   sortKey,
@@ -83,6 +94,7 @@ const GetCollection = ({
 }: {
   cms: TinaCMS
   collectionName: string
+  folder: { loading: boolean; fullyQualifiedName: string }
   includeDocuments?: boolean
   startCursor?: string
   sortKey?: string
@@ -95,6 +107,7 @@ const GetCollection = ({
       cms,
       collectionName,
       includeDocuments,
+      folder,
       startCursor || '',
       sortKey,
       filterArgs