+++ packages/tinacms/src/admin/pages/CollectionUpdatePage.tsx
@@ -15,6 +15,7 @@ import { PageWrapper } from '../components/Page'
 import { TinaAdminApi } from '../api'
 import type { TinaCMS } from '@tinacms/toolkit'
 import { useWindowWidth } from '@react-hook/window-size'
+import { useCollectionFolder } from './utils'
 
 const updateDocument = async (
   cms: TinaCMS,
@@ -37,18 +38,21 @@ const updateDocument = async (
 
 const CollectionUpdatePage = () => {
   const { collectionName, ...rest } = useParams()
-  const { '*': filename } = rest
+  const folder = useCollectionFolder()
+  const { '*': filename } = rest // TODO can just use the folder.name instead
 
+  const resolvedFile = folder.fullyQualifiedName ? folder.name : filename
   return (
     <GetCMS>
       {(cms: TinaCMS) => (
         <GetCollection
           cms={cms}
           collectionName={collectionName}
+          folder={folder}
           includeDocuments={false}
         >
           {(collection) => {
-            const relativePath = `${filename}.${collection.format}`
+            const relativePath = `${resolvedFile}.${collection.format}`
             const mutationInfo = {
               includeCollection: true,
               includeTemplate: !!collection.templates,
@@ -65,10 +69,11 @@ const CollectionUpdatePage = () => {
                     <RenderForm
                       cms={cms}
                       document={document}
-                      filename={filename}
+                      filename={resolvedFile}
                       relativePath={relativePath}
                       collection={collection}
                       mutationInfo={mutationInfo}
+                      folder={folder}
                     />
                   )}
                 </GetDocument>
@@ -88,6 +93,7 @@ const RenderForm = ({
   relativePath,
   collection,
   mutationInfo,
+  folder,
 }) => {
   const [formIsPristine, setFormIsPristine] = useState(true)
   const schema: TinaSchema | undefined = cms.api.tina.schema
@@ -147,7 +153,7 @@ const RenderForm = ({
           <div className="mb-2">
             <span className="block text-sm leading-tight uppercase text-gray-400 mb-1">
               <Link
-                to={`/collections/${collection.name}`}
+                to={`/collections/${collection.name}${folder.parentName}`}
                 className="inline-block text-current hover:text-blue-400 focus:underline focus:outline-none focus:text-blue-400 font-medium transition-colors duration-150 ease-out"
               >
                 {collection.label ? collection.label : collection.name}