+++ experimental-examples/kitchen-sink/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [709b6f2ec]
+  - tinacms@1.4.6
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [75d5ed359]
   - tinacms@1.4.5